perf(distributed): skip eager probe session when chunkWorkerCount > 1
`renderChunk` was pre-warming a `createCaptureSession + assertSwiftShader +
initializeSession` pipeline before every chunk render. When the resolved
`chunkWorkerCount > 1`, the parallel branch in `captureStage` immediately
closes that probe and spins up fresh per-worker sessions — wasting the
~3-5s of pre-warmed setup. (The OSS comment at `runCaptureStage(...
probeSession: session ...)` flagged this as a follow-up.)

Move the SwiftShader assertion into `executeWorkerTask` so each parallel
worker validates its own GPU backend against `chrome://gpu`-style canvas
probe (canvas + WEBGL_debug_renderer_info works on both regular Chrome
and `chrome-headless-shell`). Gated on `cfg.browserGpuMode === "software"`
so in-process renders that opt into software GL also pick up the safety
net, while hardware-GL paths are untouched.

In `renderChunk`, compute `chunkWorkerCount` up-front and skip the entire
pre-warmup (createCaptureSession + assert + initialize) when > 1 — the
parallel workers cover it. Sequential path (chunkWorkerCount === 1) is
unchanged: it still pre-warms because `captureStage`'s sequential branch
reuses the probe.

Move `readWebGlVendorInfoFromCanvas` from
`packages/producer/src/services/distributed/renderChunk.ts` to
`packages/engine/src/utils/readWebGlVendorInfoFromCanvas.ts` (both
producer and engine need it now). `renderChunk.ts` re-exports the
function from `@hyperframes/engine` so downstream consumers that import
it from `@hyperframes/producer/distributed` keep working (the
`publicExports.test.ts` assertion is preserved).

Expected impact on the texture-launch fixture (dev, 12 producer-worker
pods, v0.6.15 sidecar; baseline from re-run sweep with `--chunk-size 10`):

  chunks=3   chunkWorkerCount=6 → ~3-5s/chunk saved (~5-10% wall)
  chunks=6   chunkWorkerCount=3 → ~3-5s/chunk saved (~7-12% wall)
  chunks=8   chunkWorkerCount=2 → ~3-5s/chunk saved (~10-13% wall)
  chunks=12  chunkWorkerCount=1 → no change (sequential path reuses probe)
