+++ packages/engine/src/index.ts
@@ -165,6 +165,8 @@ export {
   BROWSER_GPU_NOT_SOFTWARE,
 } from "./utils/assertSwiftShader.js";
 
+export { readWebGlVendorInfoFromCanvas } from "./utils/readWebGlVendorInfoFromCanvas.js";
+
 export {
   extractMediaMetadata,
   extractVideoMetadata,
